refactor(hub-ui): DetailPanel → AgentInspector column, split tabs
Move the full-page slide-in DetailPanel to components/inspector/AgentInspector.tsx,
restyle as a column that fills the Shell inspector slot (.detail-panel--inspector),
and split the ten tab bodies into inspector/tabs/*. Content/behavior unchanged.
